Updating event work
