+++ src/constants/file.ts
@@ -1,0 +1,1 @@
+export const UNIQUE_ID_IDENTIFIER = ':upload:';
