+++ src/file-upload-with-preview.spec.js
