+++ src/utils/file.ts
@@ -1,0 +1,1 @@
+export const generateUniqueId = () => Math.random().toString(16).slice(2);
